% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrences.R
\name{build_cooccurrence_query}
\alias{build_cooccurrence_query}
\title{Build a co-occurrence query}
\usage{
build_cooccurrence_query(terms_A, terms_B)
}
\arguments{
\item{terms_A}{List of terms for entity A.}

\item{terms_B}{List of terms for entity B.}
}
\value{
A query list for Elasticsearch.
}
\description{
Build a co-occurrence query
}
