% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrences.R
\name{count_all_docs}
\alias{count_all_docs}
\title{Count documents in index}
\usage{
count_all_docs(relminer)
}
\arguments{
\item{relminer}{The RelMiner object.}
}
\value{
The count of co-occurrences.
}
\description{
Count documents in index
}
