% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrences.R
\name{count_cooccurrences}
\alias{count_cooccurrences}
\title{Count co-occurrences of two sets of terms}
\usage{
count_cooccurrences(terms_A, terms_B, index_name, relminer)
}
\arguments{
\item{terms_A}{List of terms for entity A.}

\item{terms_B}{List of terms for entity B.}

\item{index_name}{Name of the Elasticsearch index.}

\item{relminer}{The RelMiner object.}
}
\value{
The count of co-occurrences.
}
\description{
Count co-occurrences of two sets of terms
}
