% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrence.R
\name{count_occurrences}
\alias{count_occurrences}
\title{Count occurrences of a list of terms in an Elasticsearch index}
\usage{
count_occurrences(terms, index_name, relminer)
}
\arguments{
\item{terms}{List of terms to search for.}

\item{index_name}{Name of the Elasticsearch index.}

\item{relminer}{The RelMiner object.}
}
\value{
The count of occurrences.
}
\description{
Count occurrences of a list of terms in an Elasticsearch index
}
