% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{create_term}
\alias{create_term}
\title{Create a term object}
\usage{
create_term(name, syn, class)
}
\arguments{
\item{name}{The name of the term.}

\item{syn}{A list of synonyms for the term.}

\item{class}{The class/category of the term.}
}
\value{
A list representing the term.
}
\description{
Create a term object
}
