% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstracts.R
\name{find_abstracts}
\alias{find_abstracts}
\title{Retrieve abstracts containing co-occurrences of entity A and B}
\usage{
find_abstracts(A, B, index_name, relminer)
}
\arguments{
\item{A}{A term list containing the name, synonyms, and class for entity A.}

\item{B}{A term list containing the name, synonyms, and class for entity B.}

\item{index_name}{Name of the Elasticsearch index.}

\item{relminer}{The RelMiner object.}
}
\value{
A data frame of abstracts containing co-occurrences of A and B.
}
\description{
Retrieve abstracts containing co-occurrences of entity A and B
}
