% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrences.R
\name{find_rel}
\alias{find_rel}
\title{Calculate PMI and NPMI for co-occurrences}
\usage{
find_rel(A, B, relminer)
}
\arguments{
\item{A}{A term list containing the name, synonyms, and class for entity A.}

\item{B}{A term list containing the name, synonyms, and class for entity B.}

\item{relminer}{The RelMiner object.}
}
\value{
A data frame with co-occurrence counts, PMI, and NPMI values.
}
\description{
Calculate PMI and NPMI for co-occurrences
}
