% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrences.R
\name{find_rels}
\alias{find_rels}
\title{Function to run find_rel in parallel}
\usage{
find_rels(terms_a, terms_b, relminer)
}
\arguments{
\item{relminer}{The RelMiner object.}

\item{A}{A term list containing the name, synonyms, and class for entity A.}

\item{B}{A term list containing the name, synonyms, and class for entity B.}

\item{n_jobs}{the number of jobs to run in parallel.}
}
\value{
A data frame with co-occurrence counts, PMI, and NPMI values.
}
\description{
Function to run find_rel in parallel
}
