% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentences.R
\name{find_sentences_with_cooccurrences}
\alias{find_sentences_with_cooccurrences}
\title{Find sentences with co-occurrences of A and B}
\usage{
find_sentences_with_cooccurrences(A, B, index_name, relminer)
}
\arguments{
\item{A}{A term list containing the name, synonyms, and class for entity A.}

\item{B}{A term list containing the name, synonyms, and class for entity B.}

\item{index_name}{Name of the Elasticsearch index.}

\item{relminer}{The RelMiner object.}
}
\value{
A data frame with sentences containing co-occurrences of A and B.
}
\description{
Find sentences with co-occurrences of A and B
}
