% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrences.R
\name{run_cooccurrences_in_parallel}
\alias{run_cooccurrences_in_parallel}
\title{Function to run find_cooccurrences in parallel}
\usage{
run_cooccurrences_in_parallel(
  terms_a,
  terms_b,
  index_name,
  relminer,
  n_jobs = 10
)
}
\arguments{
\item{index_name}{Name of the Elasticsearch index.}

\item{relminer}{The RelMiner object.}

\item{n_jobs}{the number of jobs to run in parallel.}

\item{A}{A term list containing the name, synonyms, and class for entity A.}

\item{B}{A term list containing the name, synonyms, and class for entity B.}
}
\value{
A data frame with co-occurrence counts, PMI, and NPMI values.
}
\description{
Function to run find_cooccurrences in parallel
}
